+++ resources/views/permissions/list.blade.php
@@ -35,11 +35,15 @@
                                     {{ \Carbon\Carbon::parse($permission->created_at)->format('d M, Y') }}
                                 </td>
                                 <td class="px-6 py-3 text-center">
+                                    @can('edit permissions')
                                     <a href="{{ route('permissions.edit', $permission->id) }}"
                                         class="bg-orange-500 text-sm rounded-md text-black px-3 py-3 hover:bg-slate-50">Edit</a>
-                                    <a href="#" onclick="deletePermission({{ $permission->id }})"
+                                    @endcan
+                                    @can('delete permissions')
+                                        <a href="#" onclick="deletePermission({{ $permission->id }})"
                                         class="bg-red-500 text-sm rounded-md text-black px-3 py-3 hover:bg-slate-50">Delete</a>
-                                </td>
+                                    @endcan
+                                    </td>
                             </tr>
                         @endforeach
                     @endif
